+++ tests/CollectionManipulationTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Ramsey\Collection\Test;
 
+use Ramsey\Collection\Collection;
 use Ramsey\Collection\Exception\CollectionMismatchException;
 use Ramsey\Collection\Exception\InvalidSortOrderException;
 use Ramsey\Collection\Exception\ValueExtractionException;
@@ -214,6 +215,32 @@ class CollectionManipulationTest extends TestCase
         $this->assertEquals([$bar2, $bar1], $barCollection3->toArray());
     }
 
+    public function testdiffShouldRaiseExceptionOnDiverseCollectionType(): void
+    {
+        $barCollection = new Collection('int');
+        $fooCollection = new Collection('string');
+
+        $this->expectException(CollectionMismatchException::class);
+        $this->expectExceptionMessage(
+            'Collection items must be of type int'
+        );
+
+        $barCollection->diff($fooCollection);
+    }
+
+    public function testDiffGenericCollection(): void
+    {
+        $bar1 = new Bar(1, 'a');
+        $bar2 = new Bar(2, 'b');
+
+        $barCollection1 = new Collection(Bar::class, [$bar1]);
+        $barCollection2 = new Collection(Bar::class, [$bar1, $bar2]);
+
+        $diffCollection = $barCollection1->diff($barCollection2);
+
+        $this->assertEquals([$bar2], $diffCollection->toArray());
+    }
+
     public function testIntersectShouldRaiseExceptionOnDiverseCollections(): void
     {
         $barCollection = new BarCollection();
@@ -253,6 +280,32 @@ class CollectionManipulationTest extends TestCase
         $this->assertEquals([$bar3, $bar2, $bar1], $barCollection3->toArray());
     }
 
+    public function testIntersectShouldRaiseExceptionOnDiverseCollectionType(): void
+    {
+        $barCollection = new Collection('int');
+        $fooCollection = new Collection('string');
+
+        $this->expectException(CollectionMismatchException::class);
+        $this->expectExceptionMessage(
+            'Collection items must be of type int'
+        );
+
+        $barCollection->intersect($fooCollection);
+    }
+
+    public function testIntersectGenericCollection(): void
+    {
+        $bar1 = new Bar(1, 'a');
+        $bar2 = new Bar(2, 'b');
+
+        $barCollection1 = new Collection(Bar::class, [$bar1]);
+        $barCollection2 = new Collection(Bar::class, [$bar1, $bar2]);
+
+        $diffCollection = $barCollection1->intersect($barCollection2);
+
+        $this->assertEquals([$bar1], $diffCollection->toArray());
+    }
+
     public function testMergeShouldRaiseExceptionOnDiverseCollection(): void
     {
         $barCollection = new BarCollection();
@@ -283,5 +336,31 @@ class CollectionManipulationTest extends TestCase
         $this->assertEquals([$bar1], $barCollection1->toArray());
         $this->assertEquals([$bar2], $barCollection2->toArray());
         $this->assertEquals([$bar3], $barCollection3->toArray());
+    }
+
+    public function testMergeShouldRaiseExceptionOnDiverseCollectionType(): void
+    {
+        $barCollection = new Collection('int');
+        $fooCollection = new Collection('string');
+
+        $this->expectException(CollectionMismatchException::class);
+        $this->expectExceptionMessage(
+            'Collection items in collection with index 1 must be of type int'
+        );
+
+        $barCollection->merge($barCollection, $fooCollection);
+    }
+
+    public function testMergeGenericCollection(): void
+    {
+        $bar1 = new Bar(1, 'a');
+        $bar2 = new Bar(2, 'b');
+
+        $barCollection1 = new Collection(Bar::class, [$bar1]);
+        $barCollection2 = new Collection(Bar::class, [$bar2]);
+
+        $diffCollection = $barCollection1->merge($barCollection2);
+
+        $this->assertEquals([$bar1, $bar2], $diffCollection->toArray());
     }
 }